+++ core/src/main/java/nl/nn/adapterframework/batch/AbstractRecordHandler.java
@@ -211,7 +211,7 @@ public abstract class AbstractRecordHandler implements IRecordHandler, IWithPara
 	}
 
 	/*
-	 * this method returns a List, and therefore cannot be called 'getRecordIdentifyingFields', 
+	 * Returns a List, and therefore cannot be called 'getRecordIdentifyingFields', 
 	 * because then setRecordIdentifyingFields is not found as a setter.
 	 */  
 	public List getRecordIdentifyingFieldList() {