+++ core/src/main/java/nl/nn/adapterframework/core/IPipe.java
@@ -61,7 +61,6 @@ String getType();
 /**
   * Register a PipeForward object to this Pipe. Global Forwards are added
   * by the PipeLine. If a forward is already registered, it logs a warning.
-  * @param forward
   * @see PipeLine
   * @see PipeForward
    */