+++ core/src/main/java/nl/nn/adapterframework/pipes/RemoveFromSession.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013 Nationale-Nederlanden
+   Copyright 2013,2019 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -49,8 +49,7 @@ import org.apache.commons.lang.StringUtils;
 	}
     
 	/**
-     * checks wether the proper forward is defined.
-     * @throws ConfigurationException
+     * Checks whether the proper forward is defined.
      */
     public void configure() throws ConfigurationException {
 	    super.configure();
@@ -62,11 +61,8 @@ import org.apache.commons.lang.StringUtils;
         }
 	*/
     }
-/**
- * This is where the action takes place. Pipes may only throw a PipeRunException,
- * to be handled by the caller of this object.
- */
-public PipeRunResult doPipe(Object input, IPipeLineSession session) throws PipeRunException {
+	 
+  public PipeRunResult doPipe(Object input, IPipeLineSession session) throws PipeRunException {
 	String result = null;
 
 	String sessionKeys = getSessionKey();
@@ -94,24 +90,16 @@ public PipeRunResult doPipe(Object input, IPipeLineSession session) throws PipeR
 			}
 		}
 	}
-	
+
 	return new PipeRunResult(getForward(), result);
-}
-/**
- * The name of the key in the <code>PipeLineSession</code> to store the input in
- * {@link IPipeLineSession pipeLineSession}
- */
-public String getSessionKey() {
-	return sessionKey;
-}
-/**
- * The name of the key in the <code>PipeLineSession</code> to store the input in
- * @see IPipeLineSession
- * 
- * @param newSessionKey String
- */
-	@IbisDoc({"name of the key in the <code>pipelinesession</code> to remove. if this key is empty the input message is interpretted as key. for multiple keys use ',' as delimiter", ""})
-public void setSessionKey(String newSessionKey) {
-	sessionKey = newSessionKey;
-}
+  }
+
+	 
+	@IbisDoc({"name of the key of the entry in the <code>pipelinesession</code> to remove. if this key is empty the input message is interpretted as key. for multiple keys use ',' as delimiter", ""})
+	public void setSessionKey(String newSessionKey) {
+		sessionKey = newSessionKey;
+	}
+	public String getSessionKey() {
+		return sessionKey;
+	}
 }