+++ core/src/main/java/nl/nn/adapterframework/validation/AbstractXmlValidator.java
@@ -269,8 +269,8 @@ public abstract class AbstractXmlValidator {
 	}
 
     /**
+     * Sets schemas provider.
      * @since 5.0
-     * @param schemasProvider
      */
 	public void setSchemasProvider(SchemasProvider schemasProvider) {
 		this.schemasProvider = schemasProvider;